Allow more elements inside a label of .btn-group
Hello, 

Currently, the script doesn't allow placing other elements inside `<label>` element of a `.btn-group`. That is, for example, we can't use glyphicon as a label for a button inside a button groups like this:

```    
<div class="btn-group " data-toggle="buttons">
        <label class="btn btn-default btn-sm active">
            <input type="checkbox" checked="checked">
            <span class="glyphicon glyphicon-repeat"></span>
        </label>
</div>
```
when the span element receives a click event, the function incorrectly decides that the label is the span element itself leading to an error further on. 

This change would allow to correctly respond to an event on any direct child of a label, thus, allowing to use the said glyphicons as labels, which seems to be a popular thing. It will still throw on elements nested inside label's immediate children, but resolving for all cases might make the code inefficient, so this is a trade-off.