A quick fix inspired by twbs #22272
The use of "preventDefault()" in all cases made impossible for an input that toggled a collapse via data api to receive the click event. As a result, all checkbox toggling a collapse through this method were unable to change state, and remained unchecked (or checked depending on their initial state) on click. This fix is similar to the solution twbs uses as of today, it only "preventDefault()" for "a" tags which may change the URL on click .